requirements.txt
# VitalWatch – core
numpy>=1.24
opencv-python>=4.8
ultralytics>=8.0
mediapipe>=0.10
torch>=2.0
torchvision>=0.15

# API & dashboard
fastapi>=0.100
uvicorn[standard]>=0.22

# Optional (existing)
PyYAML>=6.0
websockets>=12.0

# database
sqlalchemy>=2.0
psycopg2-binary>=2.9
